chore: Move implementation to Windows.UI for compat
