+++ src/SamplesApp/UITests.Shared/Microsoft_UI/ColorHelperTests.xaml.cs
@@ -36,7 +36,7 @@ internal class ColorHelperTestsViewModel(UnitTestDispatcherCompat dispatcher) : 
 
 	public Command ConvertCommand => new((p) =>
 	{
-		ColorName = ColorHelper.ToDisplayName(Microsoft.UI.ColorHelper.ConvertColorFromHexString(ColorValue));
+		ColorName = ColorHelper.ToDisplayName(Windows.UI.ColorHelper.ConvertColorFromHexString(ColorValue));
 	});
 
 	public string ColorName
